# specs/grc-claw-openapi.yaml
openapi: 3.1.0
info:
  title: GRC_Claw API
  description: |
    GRC_Claw — open-source governance, risk, and compliance platform for agentic AI.
    
    This API provides four complementary interfaces:
    - **REST** — CRUD operations, search, and reporting
    - **GraphQL** — Complex nested queries for dashboards and analytics
    - **gRPC** — High-performance enforcement decisions and streaming
    - **Webhooks** — Event-driven integrations
    
    ## Authentication
    
    All API requests require authentication via one of:
    - **OAuth 2.1 / OIDC** — Bearer JWT token (RS256)
    - **API Key** — `Authorization: Bearer grc_live_...` or `grc_test_...`
    - **mTLS (SPIFFE)** — X.509 SVID for agent-to-PEP communication
    
    ## Rate Limiting
    
    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit` — Request limit per window
    - `X-RateLimit-Remaining` — Remaining requests in current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets
    
    ## Error Format
    
    All errors follow RFC 7807 (Problem Details for HTTP APIs).
    
    ## Versioning
    
    Current version: `v1.0`. Version is specified in the URL path.
    Deprecation notices are sent via webhook 6 months before sunset.
  version: 1.0.0
  contact:
    name: GRC_Claw Architecture Team
    url: https://api.grc-claw.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.grc-claw.io
...
paths:
  /v1.0/policies:
    get:
      operationId: listPolicies
      summary: List policies
      description: Returns a paginated list of policies with optional filtering.
      tags: [Policies]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, review, active, deprecated, archived]
        - name: category
          in: query
          schema:
            type: string
            enum: [ethics, safety, privacy, fairness]
        - name: framework
          in: query
          schema:
            type: string
        - name: agent_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 500
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
    post:
      operationId: createPolicy
      summary: Create policy
      description: Creates a new policy in draft status.
      tags: [Policies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyCreateRequest'
      responses:
        '201':
          description: Policy created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1.0/policies/{policy_id}:
    parameters:
      - name: policy_id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getPolicy
      summary: Get policy
      description: Returns full policy object including Cedar source and compiled Rego.
      tags: [Policies]
      responses:
        '200':
          description: Policy details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: updatePolicy
      summary: Update policy
      description: Updates an existing policy. Creates a new version.
      tags: [Policies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyUpdateRequest'
      responses:
        '200':
          description: Policy updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deletePolicy
      summary: Delete policy
      description: Deletes a policy. Use `force=true` to delete even if active.
      tags: [Policies]
      parameters:
        - name: force
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: Policy deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1.0/policies/{policy_id}/compile:
    post:
      operationId: compilePolicy
      summary: Compile policy
      description: Compiles Cedar policy to Rego for OPA evaluation.
      tags: [Policies]
      parameters:
        - name: policy_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Compilation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompilationResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Policy compilation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /v1.0/policies/{policy_id}/dry-run:
    post:
      operationId: dryRunPolicy
      summary: Dry-run policy
      description: Evaluates policy against test inputs without enforcement.
      tags: [Policies]
      parameters:
        - name: policy_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [test_inputs]
              properties:
                test_inputs:
                  type: array
                  items:
                    $ref: '#/components/schemas/DryRunInput'
      responses:
        '200':
          description: Dry-run results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DryRunResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1.0/policies/{policy_id}/versions:
    get:
      operationId: getPolicyVersions
      summary: Policy version history
      description: Returns version history for a policy.
      tags: [Policies]
      parameters:
        - name: policy_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Version history
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PolicyVersion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1.0/policies/{policy_id}/dependencies:
    get:
      operationId: getPolicyDependencies
      summary: Policy dependency graph
      description: Returns dependency graph for a policy.
      tags: [Policies]
      parameters:
        - name: policy_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dependency graph
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyDependencyGraph'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'API key in format: Bearer grc_live_... or Bearer grc_test_...'
    mtlsAuth:
      type: mutualTLS
      description: mTLS with SPIFFE SVID
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema: { type: integer }
        X-RateLimit-Remaining:
          schema: { type: integer }
        X-RateLimit-Reset:
          schema: { type: integer }
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, detail, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
        trace_id:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string
    Pagination:
      type: object
      properties:
        next_cursor:
          type: string
        has_next:
          type: boolean
        total:
          type: integer
    Policy:
      type: object
      properties:
        id: { type: string }
        policy_key: { type: string }
        name: { type: string }
        description: { type: string }
        category:
          type: string
          enum: [ethics, safety, privacy, fairness]
        status:
          type: string
          enum: [draft, review, active, deprecated, archived]
        version: { type: string }
        framework_tags:
          type: array
          items: { type: string }
        effective_date:
          type: string
          format: date-time
        expiry_date:
          type: string
          format: date-time
        owner_id: { type: string }
        agent_bindings:
          type: array
          items: { type: string }
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PolicyCreateRequest:
      type: object
      required: [policy_key, name, category]
      properties:
        policy_key: { type: string }
        name: { type: string }
        description: { type: string }
        category:
          type: string
          enum: [ethics, safety, privacy, fairness]
        framework_tags:
          type: array
          items: { type: string }
        cedar_policy: { type: string }
        metadata:
          type: object
          additionalProperties: true
    PolicyUpdateRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        cedar_policy: { type: string }
        metadata:
          type: object
          additionalProperties: true
    PolicyListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Policy'
        pagination:
          $ref: '#/components/schemas/Pagination'
    PolicyVersion:
      type: object
      properties:
        version: { type: string }
        status: { type: string }
        change_summary: { type: string }
        created_at:
          type: string
          format: date-time
        created_by: { type: string }
    PolicyDependencyGraph:
      type: object
      properties:
        policy_id: { type: string }
        dependencies:
          type: array
          items:
            type: object
            properties:
              target_policy_id: { type: string }
              target_policy_name: { type: string }
              relation_type: { type: string }
              description: { type: string }
        dependents:
          type: array
          items:
            type: object
            properties:
              source_policy_id: { type: string }
              source_policy_name: { type: string }
              relation_type: { type: string }
              description: { type: string }
    CompilationResult:
      type: object
      properties:
        policy_id: { type: string }
        compilation_status:
          type: string
          enum: [success, failed, partial]
        rego_policy: { type: string }
        warnings:
          type: array
          items: { type: string }
        errors:
          type: array
          items: { type: string }
        compiled_at:
          type: string
          format: date-time
    DryRunInput:
      type: object
      properties:
        principal:
          type: object
          additionalProperties: true
        action: { type: string }
        resource:
          type: object
          additionalProperties: true
        context:
          type: object
          additionalProperties: true
    DryRunResult:
      type: object
      properties:
        policy_id: { type: string }
        dry_run_results:
          type: array
          items:
            type: object
            properties:
              input_index: { type: integer }
              decision:
                type: string
                enum: [ALLOW, DENY]
              matched_rules:
                type: array
                items: { type: string }
              evaluation_time_ms: { type: number }
              reason: { type: string }
        summary:
          type: object
          properties:
            total: { type: integer }
            allowed: { type: integer }
            denied: { type: integer }
            avg_evaluation_time_ms: { type: number }
...